chore: add getHostPreviousSibling.js
