+++ packages/rax/src/vdom/inject.js
@@ -6,8 +6,6 @@ import CompositeComponent from './composite';
 import FragmentComponent from './fragment';
 import reconciler from '../devtools/reconciler';
 import { invokeMinifiedError } from '../error';
-import toArray from '../toArray';
-import { INTERNAL } from '../constant';
 
 export default function inject({ driver, measurer }) {
   // Inject component class
@@ -16,29 +14,6 @@ export default function inject({ driver, measurer }) {
   Host.Text = TextComponent;
   Host.Fragment = FragmentComponent;
   Host.Composite = CompositeComponent;
-
-  function getHostPreviousSibling(component) {
-    let parent = component;
-    while (parent = component._parentInstance && component._parentInstance[INTERNAL]) {
-      if (parent instanceof CompositeComponent) {
-        component = parent;
-        continue;
-      }
-
-      const keys = Object.keys(parent._renderedChildren);
-      for (let i = component.__mountIndex - 1; i >= 0; i--) {
-        const nativeNode = toArray(parent._renderedChildren[keys[i]].__getNativeNode());
-        if (nativeNode.length > 0) return nativeNode[nativeNode.length - 1];
-      }
-      if (parent instanceof FragmentComponent) {
-        component = parent;
-      } else {
-        return null;
-      }
-    }
-  }
-
-  Host.getHostPreviousSibling = getHostPreviousSibling;
 
   // Inject render driver
   if (!(Host.driver = driver || Host.driver)) {
